+++ src/Celerity.Benchmarks/Program.cs
@@ -65,6 +65,7 @@ internal class Program
         typeof(SpatialGridBenchmark),
         typeof(RTreeBenchmark),
         typeof(IntervalTreeBenchmark),
+        typeof(RangeMapBenchmark),
         typeof(CompressedGraphBenchmark),
         typeof(BTreeDictionaryBenchmark),
         typeof(BTreeSetBenchmark),